backend/requirement.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
requests
beautifulsoup4
python-dotenv
google-generativeai
